Obtain version number from git tag
Nw we have transitioned to using git tags for marking release points it makes
sense to use those tags themselves for filling in the version number in the
IDE when compiling.  Both the master build and the core build scripts use
the output of "git tag" as the version number instead of getting it from
the version.txt file.