Added ability to select to use the system dectorator
Some of the LaF themes use their own decorator.  That doesn't always
work properly on all systems as it expects things to be in certain
places.  A checkbox has now been added to the preferences to turn off
the custom decorators and use the system dectorator instead.  By default
now it will use the system dectorator unless specifically told not to.

Also in this commit is a small bug-fix to cure a NullPointerException when
updating editor menus with no core selected (first installation of any
plugins, for instance).