Added sketch-level libraries
It is now possible to place libraries in a "libraries" folder within your
sketch folder.  These libraries are used in preference to any other libraries
with the same name in the system.  In addition the archive names used when
building libraries have been expanded to include the type of library ("core",
"contributed" or "sketch") so that any override libraries in sketches don't
mess up the cached archives for other sketches.

Save As now also copies the contents of the libraries and objects folders
across to the new sketch folder.