Created basic sketch backup system
You now have the option to, when saving your sketch, to create a number
of backup copies.  The number (and the option) to save is in preferences,
and if enabled, it keeps the last N versions of your sketch files in a
"backup" folder inside the sketch.  The backups are named <sketchname>-1
through to <sketchname>-N, with -1 being the most recent copy.  At the
moment it only backs up the files that are able to be edited directly by
the IDE, so doesn't include sketch libraries, binary files, etc.