Fixed compile abort on sketch error
I bug in the compiler system meant that if the sketch (which is compiled
first in the list) has an error the compilation continued right to the end
resulting in the error being lost up the top of the output.  It now aborts
the compilation immediately the sketch compilation fails.