Embarked on a daring resign
I have started redesigning much of the system from the ground up.  Instead
of patching and tweaking the existing editor code, I decided that I should
just delete it and start from scratch.  It's much cleaner, and does a whole
load more than before.  Or it will when it's finished though.  At the moment
it's so crippled as to be pretty much unusable.  Lots and lots still to do.